requirements.txt
fastapi
flask 
uvicorn 
mysql-connector-python 
sqlalchemy
pymysql 
flask-sqlalchemy
dotenv
pydantic[dotenv] 
pydantic[email]
fastapi[all]